Working diagnostics mapping
